+++ package.json
@@ -9,16 +9,16 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "axios": "^1.4.0",
     "next": "13.2.3",
-    "typescript": "4.9.5",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "styled-components": "^6.0.5",
     "react-hook-form": "^7.43.9",
     "react-icons": "^4.8.0",
     "react-quill": "^2.0.0",
     "recoil": "^0.7.7",
-    "axios": "^1.4.0"
+    "styled-components": "^6.0.5",
+    "typescript": "4.9.5"
   },
   "devDependencies": {
     "@types/node": "18.14.6",
