+++ src/components/pages/login/LoginForm/index.tsx
@@ -1,6 +1,7 @@
 import DefaultButton from '@src/components/common/DefaultButton';
 import DefaultForm from '@src/components/common/DefaultForm';
-import DefaultInput from '@src/components/common/DefaultInput';
+import LabelInput from '@src/components/molecules/Inputs/LabelInput';
+import { PLACEHOLDER } from '@src/utils/constants';
 import { useForm } from 'react-hook-form';
 import { AiOutlineLock, AiOutlineUser } from 'react-icons/ai';
 import { styled } from 'styled-components';
@@ -13,29 +14,21 @@ interface LoginFormData {
 function LoginForm() {
   const { register } = useForm<LoginFormData>();
 
+  const idRegister = register('id', { required: true });
+  const pwRegister = register('pw', { required: true });
+
   return (
     <Form>
-      <InputWrapper>
-        <Label htmlFor="id">
-          <AiOutlineUser size={25} />
-        </Label>
-        <LoginInput
-          placeholder="ID를 입력해주세요."
-          autoComplete="off"
-          register={register('id', { required: true })}
-        />
-      </InputWrapper>
-      <InputWrapper>
-        <Label htmlFor="pw">
-          <AiOutlineLock size={25} />
-        </Label>
-        <LoginInput
-          placeholder="PW를 입력해주세요."
-          autoComplete="off"
-          type="password"
-          register={register('pw', { required: true })}
-        />
-      </InputWrapper>
+      <LabelInput
+        labelIcon={<AiOutlineUser size={25} />}
+        register={idRegister}
+        placeholder={PLACEHOLDER.id}
+      />
+      <LabelInput
+        labelIcon={<AiOutlineLock size={25} />}
+        register={pwRegister}
+        placeholder={PLACEHOLDER.pw}
+      />
       <LoginButton full onClick={() => {}}>
         로그인
       </LoginButton>
@@ -46,19 +39,7 @@ export default LoginForm;
 
 const Form = styled(DefaultForm)`
   width: 450px;
-`;
-const LoginInput = styled(DefaultInput)`
-  margin-bottom: 0.3em;
-  padding-left: 45px;
-`;
-const InputWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  width: 100%;
-`;
-const Label = styled.label`
-  position: absolute;
-  left: 15px;
+  gap: 0.5em;
 `;
 const LoginButton = styled(DefaultButton)`
   height: 50px;
